# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: zTech CRM API
  description: API del CRM de salones de eventos corporativos — UNLaM GADS1.
  version: v1
servers:
- url: http://localhost:53103
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/opportunities/{id}:
    get:
      tags:
      - Opportunities
      summary: "Detalle de una oportunidad, con cliente, salón y etapa resueltos (BE-OPP-03)"
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityDetailResponse"
        "404":
          description: No existe una oportunidad con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityDetailResponse"
    put:
      tags:
      - Opportunities
      summary: "Edita una oportunidad abierta (no cambia etapa, responsable ni estado)"
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityRequest"
        required: true
      responses:
        "200":
          description: Oportunidad actualizada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "404":
          description: No existe una oportunidad con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "409":
          description: La oportunidad está cerrada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
...
  /api/v1/opportunities:
    get:
      tags:
      - Opportunities
      summary: "Lista oportunidades visibles, paginado"
      description: SELLER ve sólo sus oportunidades; ADMIN y SALES_MANAGER ven todas
        las del tenant.
      operationId: list_1
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ABIERTA
          - GANADA
          - PERDIDA
      - name: stageId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: originId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: salesRepId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: venueId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: companyId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: contactId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: eventFrom
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: eventTo
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: "eventStart,desc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseOpportunityResponse"
    post:
      tags:
      - Opportunities
      summary: Crea una oportunidad
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityRequest"
        required: true
      responses:
        "201":
          description: Oportunidad creada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "404":
          description: "La empresa, el contacto, el salón o la etapa relacionados\
            \ no existen"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "422":
          description: "Falta empresa/contacto, o la etapa inicial no está abierta"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
  /api/v1/opportunities/{id}/stage:
    post:
      tags:
      - Opportunities
      summary: Cambia la etapa de una oportunidad abierta y registra el historial
        (BE-OPP-05)
      operationId: changeStage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeStageRequest"
        required: true
      responses:
        "200":
          description: Etapa cambiada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "404":
          description: No existe la oportunidad o la etapa destino
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "409":
          description: La oportunidad está cerrada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
        "422":
          description: "Misma etapa, o etapa destino no abierta (usar /win o /lose)"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
...
  /api/v1/opportunities/{id}/stage-history:
    get:
      tags:
      - Opportunities
      summary: Lista el historial append-only de etapas en orden cronológico
      operationId: stageHistory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StageHistoryResponse"
...
  /api/v1/opportunities/board:
    get:
      tags:
      - Opportunities
      summary: "Oportunidades visibles agrupadas por etapa, una columna por etapa\
        \ (BE-OPP-06)"
      description: SELLER ve sólo sus oportunidades; ADMIN y SALES_MANAGER ven todas
        las del tenant.
      operationId: board
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ABIERTA
          - GANADA
          - PERDIDA
      - name: stageId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: originId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: salesRepId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: venueId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: companyId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: contactId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: eventFrom
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: eventTo
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OpportunityBoardResponse"
...
components:
  schemas:
...
    UpdateOpportunityRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        companyId:
          type: integer
          format: int64
        contactId:
          type: integer
          format: int64
        venueId:
          type: integer
          format: int64
        eventTypeId:
          type: integer
          format: int64
        estimatedValue:
          type: number
        probability:
          type: integer
          format: int32
          maximum: 100
          minimum: 0
        eventStart:
          type: string
          format: date-time
        eventEnd:
          type: string
          format: date-time
        attendeeCount:
          type: integer
          format: int32
          exclusiveMinimum: 0
        estimatedCloseDate:
          type: string
          format: date
        originId:
          type: integer
          format: int64
        serviceIds:
          type: array
          items:
            type: integer
            format: int64
          uniqueItems: true
        notes:
          type: string
      required:
      - attendeeCount
      - eventEnd
      - eventStart
      - eventTypeId
      - title
      - venueId
    OpportunityResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        companyId:
          type: integer
          format: int64
        contactId:
          type: integer
          format: int64
        salesRepId:
          type: integer
          format: int64
        venueId:
          type: integer
          format: int64
        stageId:
          type: integer
          format: int64
        eventTypeId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - ABIERTA
          - GANADA
          - PERDIDA
        estimatedValue:
          type: number
        eventStart:
          type: string
          format: date-time
        eventEnd:
          type: string
          format: date-time
        attendeeCount:
          type: integer
          format: int32
        serviceIds:
          type: array
          items:
            type: integer
            format: int64
          uniqueItems: true
        companyName:
          type: string
        contactName:
          type: string
        salesRepName:
          type: string
        venueName:
          type: string
        stageName:
          type: string
...
    CreateOpportunityRequest:
      type: object
      properties:
        title:
          type: string
          example: Evento de fin de año - Empresa X
          minLength: 1
        companyId:
          type: integer
          format: int64
          description: Empresa relacionada. Al menos una de companyId/contactId es
            obligatoria.
        contactId:
          type: integer
          format: int64
          description: Contacto relacionado. Al menos una de companyId/contactId es
            obligatoria.
        salesRepId:
          type: integer
          format: int64
          description: Usuario responsable comercial (RN-01 de la consigna)
        venueId:
          type: integer
          format: int64
        stageId:
          type: integer
          format: int64
          description: Etapa inicial — debe ser una etapa abierta (DP-06)
        eventTypeId:
          type: integer
          format: int64
        estimatedValue:
          type: number
        probability:
          type: integer
          format: int32
          maximum: 100
          minimum: 0
        eventStart:
          type: string
          format: date-time
        eventEnd:
          type: string
          format: date-time
        attendeeCount:
          type: integer
          format: int32
          exclusiveMinimum: 0
        estimatedCloseDate:
          type: string
          format: date
        originId:
          type: integer
          format: int64
        serviceIds:
          type: array
          items:
            type: integer
            format: int64
          uniqueItems: true
        notes:
          type: string
      required:
      - attendeeCount
      - eventEnd
      - eventStart
      - eventTypeId
      - stageId
      - title
      - venueId
    ChangeStageRequest:
      type: object
      properties:
        stageId:
          type: integer
          format: int64
          description: Etapa destino — debe ser una etapa abierta
        note:
          type: string
          description: Observación opcional sobre el cambio
      required:
      - stageId
...
    VenueResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        capacity:
          type: integer
          format: int32
        rate:
          type: number
        address:
          type: string
        locality:
          type: string
        description:
          type: string
        equipment:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
          - DISPONIBLE
          - MANTENIMIENTO
          - INACTIVO
...
    StageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        position:
          type: integer
          format: int32
        kind:
          type: string
          enum:
          - OPEN
          - WON
          - LOST
        active:
          type: boolean
...
    PageResponseOpportunityResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OpportunityResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    CompanySummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        legalName:
          type: string
        businessName:
          type: string
        salesRepId:
          type: integer
          format: int64
    ContactSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        companyId:
          type: integer
          format: int64
        salesRepId:
          type: integer
          format: int64
    OpportunityDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        company:
          $ref: "#/components/schemas/CompanySummaryResponse"
        contact:
          $ref: "#/components/schemas/ContactSummaryResponse"
        salesRepId:
          type: integer
          format: int64
        venue:
          $ref: "#/components/schemas/VenueResponse"
        stage:
          $ref: "#/components/schemas/StageResponse"
        eventTypeId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - ABIERTA
          - GANADA
          - PERDIDA
        estimatedValue:
          type: number
        finalValue:
          type: number
        probability:
          type: integer
          format: int32
        eventStart:
          type: string
          format: date-time
        eventEnd:
          type: string
          format: date-time
        attendeeCount:
          type: integer
          format: int32
        estimatedCloseDate:
          type: string
          format: date
        closedAt:
          type: string
          format: date-time
        originId:
          type: integer
          format: int64
        lossReasonId:
          type: integer
          format: int64
        serviceIds:
          type: array
          items:
            type: integer
            format: int64
          uniqueItems: true
        notes:
          type: string
    StageHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromStage:
          $ref: "#/components/schemas/StageResponse"
        toStage:
          $ref: "#/components/schemas/StageResponse"
        changedAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/UserSummaryResponse"
        note:
          type: string
    UserSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
...
    OpportunityBoardResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            $ref: "#/components/schemas/StageColumn"
    StageColumn:
      type: object
      properties:
        stageId:
          type: integer
          format: int64
        stageName:
          type: string
        opportunities:
          type: array
          items:
            $ref: "#/components/schemas/OpportunityResponse"
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT